# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RSS Discovery API",
    "version": "1.0.0",
    "description": "Public feed catalog, search, discovery, and full-text cache API. Authenticate with Bearer tokens (rd_<level>_<secret>). Levels: readonly, standard, priority, admin."
  },
  "servers": [{"url": "/", "description": "This host"}],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "rd_<level>_<secret>"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
  },
  "security": [{"bearerAuth": []}, {"apiKeyHeader": []}],
  "paths": {
...
    "/v1/feeds": {
      "get": {
        "summary": "List feeds",
        "parameters": [
          {"name": "limit", "in": "query", "schema": {"type": "integer"}},
          {"name": "offset", "in": "query", "schema": {"type": "integer"}},
          {"name": "category", "in": "query", "schema": {"type": "string"}}
        ],
        "responses": {"200": {"description": "Feed list"}}
      },
      "post": {
        "summary": "Add a feed URL",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url"],
                "properties": {
                  "url": {"type": "string", "format": "uri"},
                  "title": {"type": "string"},
                  "category": {"type": "string"},
                  "blurb": {"type": "string"},
                  "site_url": {"type": "string"},
                  "refresh": {"type": "boolean"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Created"},
          "400": {"description": "Invalid URL"}
        }
      }
    },
...
    "/v1/search/feeds": {
      "get": {
        "summary": "Search feeds",
        "parameters": [
          {"name": "q", "in": "query", "required": true, "schema": {"type": "string"}},
          {"name": "limit", "in": "query", "schema": {"type": "integer"}}
        ],
        "responses": {"200": {"description": "Matches"}}
      }
    },
...
  }
}